backend/pyproject.toml
[project]
name = "interview-coach-backend"
version = "0.1.0"
description = "Backend for Real-Time Interview Coach"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "vision-agents[getstream,openai,elevenlabs,deepgram]",
    "fastapi",
    "uvicorn",
    "python-dotenv"
]

[tool.ruff]
line-length = 88
